ir: Move Stencil to ir/support
