+++ devito/dse/aliases.py
@@ -4,9 +4,9 @@ from collections import OrderedDict, namedtuple
 import numpy as np
 from sympy import Indexed
 
+from devito.ir.support import Stencil
 from devito.exceptions import DSEException
 from devito.symbolics import retrieve_indexed, q_indirect
-from devito.stencil import Stencil
 
 __all__ = ['collect']
