+++ devito/ir/support/__init__.py
@@ -1,1 +1,2 @@
 from devito.ir.support.basic import *  # noqa
+from devito.ir.support.stencil import *  # noqa
