+++ devito/ir/support/stencil.py
