+++ tests/test_dse.py
@@ -6,11 +6,10 @@ import pytest
 from conftest import x, y, z, time, skipif_yask  # noqa
 
 from devito import Eq  # noqa
-from devito.ir import Expression, FindNodes, clusterize, temporaries_graph
+from devito.ir import Stencil, Expression, FindNodes, clusterize, temporaries_graph
 from devito.dse import rewrite, common_subexprs_elimination, collect
 from devito.symbolics import (xreplace_constrained, iq_timeinvariant, iq_timevarying,
                               estimate_cost, pow_to_mul)
-from devito.stencil import Stencil
 from devito.types import Scalar
 from examples.seismic.acoustic import AcousticWaveSolver
 from examples.seismic import demo_model, RickerSource, GaborSource, Receiver
